DataUniverse/backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
pgvector
pandas
openpyxl
boto3
pyarrow
pyyaml
dask[complete]
distributed
bokeh
